redone hamburger menu
